Remove string interning in facts for now.
Value::String types are not safe to intern for the lifetime of the
program, because their range is unbounded; there is no limit on the
number of unique string values that could be used in a query, for
instance.

I left interning in place for logic variables for now. It might not
actually be safe either to intern those statically; better would be
to intern them per-query, or something along those lines.

It *would* be safe (and desirable) to intern attribute names. They
are currently represented as a Value::String, which makes that impossible,
but they could easily be represented with a different type that is
interned. I'm leaving the string_ref module in place primarily for that
reason.