Finish separate transactor.
Heap transactor handling is a little messy, but that's not a big deal.

As a side effect of implementing the separate transactor, the conn/db duality
has been made into two separate structs, Conn and Db, where Conn represents the
connection to the database (where you can perform transactions or retrieve a reference
to the DB at a point in time) and Db represents a queryable reference into the db at
a point in time.