Make inserting facts into the DB idempotent (#38)
Logically, this change makes it so that adding a fact which is already
in the database (i.e. matches a current fact in entity, attribute,
value, and addition/retraction status) does not modify the db.

In order to avoid checking the durable store every time we add a fact
to the db, this is implemented by allowing duplicates in the in-memory
tree, and deduplicating them when reading the index and when
rebuilding the durable index so that the dupes don't get persisted to storage.