Switch back to 0MQ for networking
This allows many simplifications including unifying the local and
networked transactor setups.

This commit includes assorted simplifications along the way such as
updated serdes. Tests remain in something of a kludgy state -- it
would be better to refactor so that Conn (or a derivative) can run a
local transactor and we can avoid the whole with_test_conn! mess.

Probably still some deadlocks here in the threading situation. That
could use a thorough review and architecture diagram.