Add indexed attribute
Also required adding boolean type. Allows us to use the AVET index
when an attribute is indexed.

Of course, right now they're actually all indexed. In the future, we
need to modify the transactor to only update the AVET index with
indexed values, and to asynchronously update the schema when an
indexed value becomes indexed.